Refactor message subsystem into modular components
